pyproject.toml
[project]
name = "cgpclient"
version = "0.1.0"
description = ""
authors = [
    {name = "Graham Ritchie",email = "gritchie@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pydantic (>=2.10.6,<3.0.0)",
    "boto3>=1.34.36",
    "requests>=2.31.0",
    "pyjwt>=2.10.0",
    "cryptography>=43.0.3",
    "pyyaml>=6.0.2",
    "fhir.resources==7.1.0"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
pytest-cov = "^5.0.0"
pytest-xdist = "^3.6.1"

[tool.pytest.ini_options]
addopts = "-vv -ra --cache-clear -p no:cacheprovider -p no:ddtrace --color=yes --cov --cov-report term --cov-report xml --junitxml=report.xml --numprocesses=auto"
testpaths = ["tests"]

[tool.coverage.run]
source = ["cgpclient"]
branch = true

[tool.coverage.report]
fail_under = 50
precision = 2
skip_empty = true

[tool.coverage.xml]
output = "coverage.xml"

[tool.isort]
profile = "black"

[tool.ruff]
exclude = [
    ".git",
    ".ruff_cache",
    ".venv",
    "__pypackages__",
    "dist",
    "venv",
]
line-length = 88

[tool.ruff.lint]
ignore = [
    "E501", # prevent line-length violations
]

[tool.ruff.per-file-ignores]
"__init__.py" = [
    "F401", # prevent cleaning up unused imports
    "F403", # allow undefined local import wildcard
]
